backend/requirements/base.txt
Django==5.1.5
psycopg[binary]==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
pgvector==0.3.6
celery==5.4.0
redis==5.2.1
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
pypdf==5.2.0
whitenoise==6.8.2
